+++ Prowl.Runtime/Audio/Null/NullAudioSource.cs
@@ -9,16 +9,16 @@ public class NullAudioSource : ActiveAudio
 {
     public override Double3 Direction { get; set; }
 
-    public override float Gain { get; set; }
-    public override float Pitch { get; set; }
-    public override float MaxDistance { get; set; }
+    public override double Gain { get; set; }
+    public override double Pitch { get; set; }
+    public override double MaxDistance { get; set; }
     public override bool Looping { get; set; }
 
     public override Double3 Position { get; set; }
 
     public override AudioPositionKind PositionKind { get; set; }
 
-    public override float PlaybackPosition { get { return 1f; } set { } }
+    public override double PlaybackPosition { get { return 1f; } set { } }
 
     public override bool IsPlaying => false;
